requirements.txt
google-genai
python-dotenv
fade
colorama
